Consolidate broken comm testing utilities
